Added cartStore, renaming state folder to store
